Replace totally the previous box
