+++ pages/index.css
@@ -1,5 +1,3 @@
-@import url(./box-theme.generated.css);
-
 html {
   font-family: Roboto, Arial, Helvetica, sans-serif;
   font-weight: 400;
