+++ pages/theme.ts
@@ -1,39 +1,6 @@
-import Box2 from '../src/box2';
-import Theme from '../src/core/theme';
+import Box from '../src/box';
 
-export const themeProps = {
-  colors: {
-    white: '#fff',
-    black: '#07071b',
-    black1: '#1e293b',
-    violet: '#988bee',
-    violetLight: '#e8edfd',
-    violetLighter: '#f6f8fe',
-    violetDark: '#5f3e66', // '#4a324f',
-    gray1: '#94a3b833',
-    gray2: '#94a3b8',
-    dark: '#272822',
-    red: 'red',
-  },
-  backgrounds: {
-    stripes: 'linear-gradient(135deg,var(--colorviolet) 10%,#0000 0,#0000 50%,var(--colorviolet) 0,var(--colorviolet) 60%,#0000 0,#0000);',
-    bg: 'url(img/bg.jpg)',
-    bg2: 'background: linear-gradient(rgba(255,255,255,.5), rgba(255,255,255,.5)), url(img/bg.jpg)',
-  },
-  backgroundImages: {
-    bg: 'linear-gradient(19deg, white 80%, rgba(183, 33, 255, 0.05) 94%)',
-    check:
-      "url(\"data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='100%' viewBox='0 0 20 20'%3e%3cpath fill='none' stroke='%23FFF' stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='M6 10l3 3l6-6'/%3e%3c/svg%3e\")",
-    radio:
-      "url(\"data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='100%' viewBox='0 0 20 20'%3e%3ccircle fill='%23FFF' cx='10' cy='10' r='5'/%3e%3c/svg%3e\")",
-    radio2:
-      "url(\"data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='100%' viewBox='0 0 20 20'%3e%3ccircle fill='%23FFF' cx='10' cy='10' r='5'/%3e%3c/svg%3e\")",
-    indeterminate:
-      "url(\"data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='100%' viewBox='0 0 20 20'%3e%3cline stroke='%23988bee' x1='4' y1='10' x2='16' y2='10' stroke-width='1' /%3e%3c/svg%3e\")",
-  },
-};
-
-Theme.setup({
+Box.themeSetup({
   button: {
     styles: {
       display: 'inline-flex',
@@ -81,10 +48,10 @@ Theme.setup({
       },
       checked: {
         bgColor: 'violet',
-        backgroundImage: 'check',
+        backgroundImage: 'bgImgCheck',
       },
       indeterminate: {
-        backgroundImage: 'indeterminate',
+        backgroundImage: 'bgImgIndeterminate',
       },
       disabled: {
         cursor: 'not-allowed',
@@ -106,7 +73,7 @@ Theme.setup({
       },
       checked: {
         bgColor: 'violet',
-        backgroundImage: 'radio',
+        backgroundImage: 'bgImgRadio',
       },
       disabled: {
         cursor: 'not-allowed',
@@ -163,10 +130,12 @@ Theme.setup({
   },
   label: {
     styles: {
-      hasInvalid: {
-        color: 'red',
-        transition: 'none',
-      },
+      // has: {
+      //   invalid: {
+      //     color: 'red',
+      //     transition: 'none',
+      //   },
+      // },
     },
   },
   components: {
@@ -175,18 +144,18 @@ Theme.setup({
         p: 3,
         b: 1,
       },
-      children: {
-        item1: {
-          styles: {
-            textTransform: 'uppercase',
-          },
-        },
-        item2: {
-          styles: {
-            textDecoration: 'underline',
-          },
-        },
-      },
+      // children: {
+      //   item1: {
+      //     styles: {
+      //       textTransform: 'uppercase',
+      //     },
+      //   },
+      //   item2: {
+      //     styles: {
+      //       textDecoration: 'underline',
+      //     },
+      //   },
+      // },
     },
     code: {
       styles: {
@@ -225,7 +194,7 @@ const colors = [
   'red',
 ] as const;
 
-export const extendedProps = Box2.extend(
+export const extendedProps = Box.extend(
   {
     white: '#fff',
     black: '#07071b',
@@ -238,6 +207,16 @@ export const extendedProps = Box2.extend(
     gray2: '#94a3b8',
     dark: '#272822',
     red: 'red',
+    bgImg1: 'linear-gradient(19deg, white 80%, rgba(183, 33, 255, 0.05) 94%)',
+    bgImgCheck:
+      "url(\"data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='100%' viewBox='0 0 20 20'%3e%3cpath fill='none' stroke='%23FFF' stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='M6 10l3 3l6-6'/%3e%3c/svg%3e\")",
+    bgImgRadio:
+      "url(\"data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='100%' viewBox='0 0 20 20'%3e%3ccircle fill='%23FFF' cx='10' cy='10' r='5'/%3e%3c/svg%3e\")",
+    bgImgRadio2:
+      "url(\"data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='100%' viewBox='0 0 20 20'%3e%3ccircle fill='%23FFF' cx='10' cy='10' r='5'/%3e%3c/svg%3e\")",
+    bgImgIndeterminate:
+      "url(\"data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' width='100%' viewBox='0 0 20 20'%3e%3cline stroke='%23988bee' x1='4' y1='10' x2='16' y2='10' stroke-width='1' /%3e%3c/svg%3e\")",
+    stripes: 'linear-gradient(135deg,var(--violet) 10%,#0000 0,#0000 50%,var(--violet) 0,var(--violet) 60%,#0000 0,#0000)',
   },
   {
     color: [
@@ -267,5 +246,20 @@ export const extendedProps = Box2.extend(
         styleName: 'outline-color',
       },
     ],
+    background: [
+      {
+        values: ['stripes'] as const,
+        valueFormat: (value: string) => `var(--${value})`,
+      },
+    ],
+    backgroundImage: [
+      {
+        values: ['bgImg1', 'bgImgCheck', 'bgImgRadio', 'bgImgRadio2', 'bgImgIndeterminate'] as const,
+        valueFormat: (value: string) => `var(--${value})`,
+        styleName: 'background-image',
+      },
+    ],
+    fill: [{ values: colors, valueFormat: (value: string) => `var(--${value})` }],
+    stroke: [{ values: colors, valueFormat: (value: string) => `var(--${value})` }],
   },
 );
