+++ src/components/flex.tsx
@@ -1,13 +1,6 @@
 import { Ref, forwardRef, RefAttributes } from 'react';
 import Box from '../box';
-
-type ExtractElementType<T> = T extends React.DetailedHTMLProps<React.HTMLAttributes<infer E>, infer E>
-  ? E
-  : T extends React.SVGProps<infer E>
-  ? E
-  : never;
-
-type ExtractElementFromTag<T extends keyof React.JSX.IntrinsicElements> = ExtractElementType<React.JSX.IntrinsicElements[T]>;
+import { ExtractElementFromTag } from '../core/coreTypes';
 
 type BoxProps<TTag extends keyof React.JSX.IntrinsicElements = 'div'> = Omit<React.ComponentProps<typeof Box<TTag>>, 'ref'>;
