--- src/core/theme2.ts
@@ -1,54 +1,0 @@
-import { useMemo } from 'react';
-import { BoxThemeStyles, ThemeProps } from '../types';
-
-export interface ThemeComponentStyles {
-  styles: BoxThemeStyles;
-  themes?: {
-    [name: string]: BoxThemeStyles;
-  };
-  // children?: {
-  //   [name: string]: ThemeComponentStyles;
-  // };
-}
-
-export interface ThemeSetup {
-  components?: {
-    [name: string]: ThemeComponentStyles;
-  };
-  button?: ThemeComponentStyles;
-  textbox?: ThemeComponentStyles;
-  textarea?: ThemeComponentStyles;
-  checkbox?: ThemeComponentStyles;
-  radioButton?: ThemeComponentStyles;
-  label?: ThemeComponentStyles;
-}
-
-export function useTheme(props: ThemeProps): BoxThemeStyles | undefined {
-  const { clean, theme, component } = props;
-
-  return useMemo(() => {
-    if (clean) return undefined;
-
-    let componentStyles = ThemeInternal.components?.[component as keyof ThemeSetup['components']] as ThemeComponentStyles;
-    if (!componentStyles) return undefined;
-
-    return theme ? { ...componentStyles.styles, ...componentStyles.themes?.[theme] } : componentStyles.styles;
-  }, [component, clean, theme]);
-}
-
-namespace ThemeInternal {
-  export let components: { [name: string]: ThemeComponentStyles } = {};
-}
-
-namespace Theme {
-  export function setup(styles: ThemeSetup) {
-    const { components, ...restStyles } = styles;
-    const com = components ?? {};
-
-    Object.entries(restStyles).forEach(([name, componentStructure]) => {
-      com[name] = componentStructure;
-    });
-
-    ThemeInternal.components = com;
-  }
-}
