+++ vite.config.ts
@@ -12,8 +12,6 @@ const componentsEntry = files.reduce((acc, fileName) => {
 
 const entry = {
   box: path.resolve(__dirname, './src/box.ts'),
-  box2: path.resolve(__dirname, './src/box2.ts'),
-  'core/theme': path.resolve(__dirname, './src/core/theme.ts'),
   ssg: path.resolve(__dirname, './src/ssg.ts'),
   ...componentsEntry,
 };
@@ -43,14 +41,6 @@ export default defineConfig(({ mode }) => {
         output: {
           inlineDynamicImports: false,
           manualChunks(id: string) {
-            if (id.includes('/core/')) {
-              return 'core';
-            }
-
-            if (id.includes('/utils/')) {
-              return 'utils';
-            }
-
             if (id.includes('/components/')) {
               const re = new RegExp('(.*)src/components/(.*)');
               const result = re.exec(id)[2];
@@ -66,9 +56,11 @@ export default defineConfig(({ mode }) => {
               return 'box';
             }
 
-            if (id.endsWith('src/box2.ts')) {
-              return 'box2';
+            if (id.endsWith('src/ssg.ts')) {
+              return 'ssg';
             }
+
+            return 'core';
           },
           chunkFileNames: (info) => `[name].${extensions[currentFormat]}`,
         },
